WIP nav and safeArgs
