+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -2,16 +2,8 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/main_nav_graph"
-    app:startDestination="@id/mainFragment">
+    app:startDestination="@id/usersListFragment">
 
-    <fragment
-        android:id="@+id/mainFragment"
-        android:name="com.example.users.ui.MainFragment"
-        android:label="MainFragment" >
-        <action
-            android:id="@+id/action_mainFragment_to_userDetailsFragment"
-            app:destination="@id/userDetailsFragment" />
-    </fragment>
     <fragment
         android:id="@+id/userDetailsFragment"
         android:name="com.example.users.ui.UserDetailsFragment"
@@ -22,5 +14,13 @@
         <argument
             android:name="userId"
             app:argType="integer" />
+    </fragment>
+    <fragment
+        android:id="@+id/usersListFragment"
+        android:name="com.example.users.ui.UsersListFragment"
+        android:label="UsersListFragment" >
+        <action
+            android:id="@+id/action_usersListFragment_to_userDetailsFragment"
+            app:destination="@id/userDetailsFragment" />
     </fragment>
 </navigation>
